Refactor folder structure to add Screens folder
