+++ client/client-app/screens/Login.js
@@ -1,6 +1,6 @@
 import { StatusBar } from "expo-status-bar";
 import React, { useState } from "react";
-import styles from "./styles";
+import styles from "../styles";
 
 import {
   StyleSheet,
